split old network train tests to trainer / validate
