+++ test/Network_test.js
@@ -86,167 +86,27 @@ describe('Network', () => {
     })
   })
 
-  describe('train', () => {
-    it('calls train on the input layer', () => {
-      network.inputLayer.train = sandbox.spy()
-      network.inputLayer.train.called.should.equal(false)
-      network.train()
-      network.inputLayer.train.called.should.equal(true)
+  describe('backprop', () => {
+    it('calls backprop on the output layer', () => {
+      network.outputLayer.backprop = sandbox.spy()
+      network.outputLayer.backprop.called.should.equal(false)
+      network.backprop()
+      network.outputLayer.backprop.called.should.equal(true)
     })
 
-    it('calls train on the input layer', () => {
+    it('calls backprop on each hidden layer', () => {
       network = new Network([2, 2, 1])
-      _.each(network.hiddenLayers, l => l.train = sandbox.spy())
-      _.each(network.hiddenLayers, l => l.train.called.should.equal(false))
-      network.train()
-      _.each(network.hiddenLayers, l => l.train.called.should.equal(true))
+      _.each(network.hiddenLayers, l => l.backprop = sandbox.spy())
+      _.each(network.hiddenLayers, l => l.backprop.called.should.equal(false))
+      network.backprop()
+      _.each(network.hiddenLayers, l => l.backprop.called.should.equal(true))
     })
 
-    it('calls train on the output layer', () => {
-      network.outputLayer.train = sandbox.spy()
-      network.outputLayer.train.called.should.equal(false)
-      network.train()
-      network.outputLayer.train.called.should.equal(true)
-    })
-  })
-
-  describe('train', () => {
-    it('is a function', () => {
-      network.train.should.be.a('function')
-    })
-
-    describe('options', () => {
-      const misuse = badOptions => network.train(data, badOptions)
-
-      describe('validation', () => {
-        it('throws if "errorThreshold" is not a number', () => {
-          expect(_.partial(misuse, {errorThreshold: ''})).to.throw()
-        })
-
-        it('throws if "frequency" is not a number', () => {
-          expect(_.partial(misuse, {frequency: ''})).to.throw()
-        })
-
-        it('throws if "maxEpochs" is not a number', () => {
-          expect(_.partial(misuse, {maxEpochs: ''})).to.throw()
-        })
-
-        it('throws if "onFail" is not a function', () => {
-          expect(_.partial(misuse, {onFail: ''})).to.throw()
-        })
-
-        it('throws if "onProgress" is not a function', () => {
-          expect(_.partial(misuse, {onProgress: ''})).to.throw()
-        })
-
-        it('throws if "onSuccess" is not a function', () => {
-          expect(_.partial(misuse, {onSuccess: ''})).to.throw()
-        })
-      })
-
-      describe('errorThreshold', () => {
-        it('controls how low the "error" must become before succeeding', () => {
-          // Infinity means we'll always have instant training success
-          network.train(data, {
-            errorThreshold: Infinity,
-            onSuccess: (error, epoch) => epoch.should.equal(1),
-          })
-        })
-      })
-
-      describe('frequency', () => {
-        it('controls how often "onProgress" is called', () => {
-          const frequency = _.random(1, 10)
-          let counter = 1
-          network.train(data, {
-            maxEpochs: 100,
-            frequency,
-            onProgress: (error, epoch) => {
-              epoch.should.equal(frequency * counter)
-              counter += 1
-            },
-          })
-        })
-      })
-
-      describe('maxEpochs', () => {
-        it('controls long the network trains for', () => {
-          const maxEpochs = _.random(1, 100)
-          // cannot solve XOR, would run forever
-          network = new Network([2, 1]).train(data, {
-            maxEpochs,
-            onFail: (error, epoch) => epoch.should.equal(maxEpochs),
-          })
-        })
-      })
-
-      describe('onFail', () => {
-        it('is called when max epochs is reached', () => {
-          network.train(data, {
-            maxEpochs: 1,
-            onFail: (error, epoch) => epoch.should.equal(1),
-          })
-        })
-
-        it('is called with "error" and "epoch" values', () => {
-          network.train(data, {
-            maxEpochs: 1,
-            onFail: (error, epoch) => {
-              epoch.should.be.a('number')
-              epoch.should.equal(1)
-              error.should.be.a('number')
-            },
-          })
-        })
-      })
-
-      describe('onProgress', () => {
-        it('is called with "error" and "epoch" values', () => {
-          network.train(data, {
-            maxEpochs: 1,
-            onProgress: (error, epoch) => {
-              epoch.should.be.a('number')
-              epoch.should.equal(1)
-              error.should.be.a('number')
-            },
-          })
-        })
-
-        it('stops training when false is returned', () => {
-          // cannot solve XOR, would run run forever
-          network = new Network([2, 1])
-          network.train(DATA.XORGate, {
-            // ensure we run a few epochs, calling progress every epoch
-            maxEpochs: 5,
-            frequency: 1,
-            onProgress: (error, epoch) => {
-              // we should only hit the first epoch
-              epoch.should.equal(1)
-              return false
-            },
-          })
-        })
-      })
-
-      describe('onSuccess', () => {
-        it('is called when "error" falls below "errorThreshold"', () => {
-          network.train(data, {
-            errorThreshold: Infinity,
-            onSuccess: (error, epoch) => epoch.should.equal(1),
-          })
-        })
-
-        it('is called with "error" and "epoch" values', () => {
-          network.train(data, {
-            maxEpochs: 1,
-            onSuccess: (error, epoch) => {
-              epoch.should.be.a('number')
-              epoch.should.equal(1)
-              error.should.be.a('number')
-            },
-          })
-        })
-      })
+    it('calls backprop on the input layer', () => {
+      network.inputLayer.backprop = sandbox.spy()
+      network.inputLayer.backprop.called.should.equal(false)
+      network.backprop()
+      network.inputLayer.backprop.called.should.equal(true)
     })
   })
 })
